requirements.txt
Django==3.2.4
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
Pillow==10.4.0
django-filter==23.1




django-debug-toolbar==3.8.1
